+++ Tests/XMTPTests/MessageTests.swift
@@ -10,7 +10,7 @@ import XCTest
 
 @available(iOS 16.0, *)
 class MessageTests: XCTestCase {
-	func testFullyEncodesDecodesMessages() async throws {
+	func testFullyEncodesDecodesMessagesV1() async throws {
 		let aliceWallet = try PrivateKey.generate()
 		let bobWallet = try PrivateKey.generate()
 
@@ -38,9 +38,31 @@ class MessageTests: XCTestCase {
 		XCTAssertEqual(message2Decrypted, content)
 	}
 
-	func testCanDecrypt() throws {
-		let content = "0a120a08786d74702e6f7267120474657874180112110a08656e636f64696e6712055554462d3822026869".web3.bytesFromHex!
+	func testFullyEncodesDecodesMessagesV2() async throws {
+		let aliceWallet = try PrivateKey.generate()
+		let bobWallet = try PrivateKey.generate()
 
+		let alice = try await PrivateKeyBundleV1.generate(wallet: aliceWallet)
+		let bob = try await PrivateKeyBundleV1.generate(wallet: bobWallet)
+
+		let client = try await Client.create(account: aliceWallet)
+		var invitationContext = InvitationV1.Context()
+		invitationContext.conversationID = "https://example.com/1"
+
+		let invitationv1 = try InvitationV1.createRandom(context: invitationContext)
+		let content = Data("Yo!".utf8)
+
+		let sealedInvitation = try SealedInvitation.createV1(sender: alice.toV2(), recipient: bob.toV2().getPublicKeyBundle(), created: Date(), invitation: invitationv1)
+		let conversation = try ConversationV2.create(client: client, invitation: invitationv1, header: sealedInvitation.v1.header)
+		let message1 = try await MessageV2.encode(client: client, content: "Yo!", topic: invitationv1.topic, keyMaterial: invitationv1.aes256GcmHkdfSha256.keyMaterial)
+
+		let decoded = try MessageV2.decode(message1, keyMaterial: invitationv1.aes256GcmHkdfSha256.keyMaterial)
+		XCTAssertEqual(decoded.body, "Yo!")
+	}
+
+	func testCanDecrypt() throws {
+		// All of these values were generated from xmtp-js
+		let content = "0a120a08786d74702e6f7267120474657874180112110a08656e636f64696e6712055554462d3822026869".web3.bytesFromHex!
 		let salt = "48c6c40ce9998a8684937b2bd90c492cef66c9cd92b4a30a4f811b43fd0aed79".web3.bytesFromHex!
 		let nonce = "31f78d2c989a37d8471a5d40".web3.bytesFromHex!
 		let secret = "04c86317929a0c223f44827dcf1290012b5e6538a54282beac85c2b16062fc8f781b52bea90e8c7c028254c6ba57ac144a56f054d569c340e73c6ff37aee4e68fc04a0fdb4e9c404f5d246a9fe2308f950f8374b0696dd98cc1c97fcbdbc54383ac862abee69c107723e1aa809cfbc587253b943476dc89c126af4f6515161a826ca04801742d6c45ee150a28f80cbcffd78a0210fe73ffdd74e4af8fd6307fb3d622d873653ca4bd47deb4711ef02611e5d64b4bcefcc481e236979af2b6156863e68".web3.bytesFromHex!
